stream-mic-to-mqtt/pyproject.toml
[project]
name = "stream-mic-to-mqtt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt",
    "pyaudio"
]
